Add timer page
This adds a new page to the app that can be used as a timer for solving.

The timer can be started and stopped with the spacebar or by click on the screen so that works well on both desktop and mobile.

This page can also be easily refactored into a reusable timer component that can be used for algorithm training.

This also adds npm caching to improve CI speed, a Playwright extension, parallelizes tests in CI, and adds Tailwind.